# frontend/content-hub-frontend/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /login/updateKakaoLoginInfo:
    get:
      tags:
      - login-controller
      operationId: updateKakaoLoginInfo
      parameters:
      - name: client_id
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginUserResponseDto"
  /login/getNaverUserInfo:
    get:
      tags:
      - login-controller
      operationId: getNaverUserInfo
      parameters:
      - name: access_token
        in: query
        required: true
        schema:
          type: string
      - name: expires_in
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginUserResponseDto"
  /login/getNaverLoginInfo:
    get:
      tags:
      - login-controller
      operationId: getNaverLoginInfo
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      - name: state
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginUserResponseDto"
  /login/getKakaoUserInfo:
    get:
      tags:
      - login-controller
      operationId: getKakaoUserInfo
      parameters:
      - name: access_token
        in: query
        required: true
        schema:
          type: string
      - name: expires_in
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginUserResponseDto"
  /login/getKakaoLoginInfo:
    get:
      tags:
      - login-controller
      operationId: getKakaoLoginInfo
      parameters:
      - name: client_id
        in: query
        required: true
        schema:
          type: string
      - name: redirect_uri
        in: query
        required: true
        schema:
          type: string
      - name: code
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginUserResponseDto"
  /login/deleteNaverToken:
    get:
      tags:
      - login-controller
      operationId: deleteNaverToken
      parameters:
      - name: access_token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NaverDeleteTokenDto"
  /login/deleteKakaoToken:
    get:
      tags:
      - login-controller
      operationId: deleteKakaoToken
      parameters:
      - name: access_token
        in: query
        required: true
        schema:
          type: string
      - name: target_id
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KakaoUserInfoDto"
...
components:
  schemas:
...
    LoginUserInfoDto:
      type: object
      properties:
        id:
          type: string
        nickname:
          type: string
        name:
          type: string
        email:
          type: string
        gender:
          type: string
        age:
          type: string
        birthday:
          type: string
        profileImage:
          type: string
        birthyear:
          type: string
        mobile:
          type: string
    LoginUserResponseDto:
      type: object
      properties:
        resultcode:
          type: string
        message:
          type: string
        userInfo:
          $ref: "#/components/schemas/LoginUserInfoDto"
        accessToken:
          type: string
        jwt:
          type: string
        expiresIn:
          type: integer
          format: int32
        expireDate:
          type: string
    NaverDeleteTokenDto:
      type: object
      properties:
        accessToken:
          type: string
        result:
          type: string
        expiresIn:
          type: integer
          format: int32
        error:
          type: string
        errorDescription:
          type: string
    KakaoAccountDto:
      type: object
      properties:
        profileNeedsAgreement:
          type: boolean
        profileNicknameNeedsAgreement:
          type: boolean
        profileImageNeedsAgreement:
          type: boolean
        profile:
          $ref: "#/components/schemas/KakaoProfileDto"
        nameNeedsAgreement:
          type: boolean
        name:
          type: string
        emailNeedsAgreement:
          type: boolean
        email:
          type: string
        ageRangeNeedsAgreement:
          type: boolean
        ageRange:
          type: string
        birthyearNeedsAgreement:
          type: boolean
        birthyear:
          type: string
        birthdayNeedsAgreement:
          type: boolean
        birthday:
          type: string
        birthdayType:
          type: string
        genderNeedsAgreement:
          type: boolean
        gender:
          type: string
        phoneNumberNeedsAgreement:
          type: boolean
        phoneNumber:
          type: string
        ciNeedsAgreement:
          type: boolean
        ci:
          type: string
        ciAuthenticatedAt:
          type: string
          format: date-time
        emailVerified:
          type: boolean
        emailValid:
          type: boolean
        leapMonth:
          type: boolean
    KakaoPartnerDto:
      type: object
      properties:
        uuid:
          type: string
    KakaoProfileDto:
      type: object
      properties:
        nickname:
          type: string
        thumbnailImageUrl:
          type: string
        profileImageUrl:
          type: string
        defaultNickname:
          type: boolean
        defaultImage:
          type: boolean
    KakaoUserInfoDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        hasSignedUp:
          type: boolean
        connectedAt:
          type: string
          format: date-time
        synchedAt:
          type: string
          format: date-time
        properties:
          type: object
          additionalProperties:
            type: string
        kakaoAccount:
          $ref: "#/components/schemas/KakaoAccountDto"
        forPartner:
          $ref: "#/components/schemas/KakaoPartnerDto"
...